fix: coords and spectral data

+++ KaydenMiller.BattleTech.Helper.Cli/Program.cs
@@ -8,16 +8,16 @@ using Microsoft.Playwright;
 using var playwright = await Playwright.CreateAsync();
 var chrome = await playwright.Chromium.LaunchAsync(new BrowserTypeLaunchOptions()
 {
-    Headless = false 
+    Headless = false
 });
 
 var page = await chrome.NewPageAsync();
 
-var s = await page.SearchSolarSystem(new KaydenMiller.BattleTech.Helper.Cli.System()
-{
-    Name = "Monument to Man",
-    SystemHref = "/wiki/Monument%20to%20Man"
-});
+// var s = await page.SearchSolarSystem(new KaydenMiller.BattleTech.Helper.Cli.System()
+// {
+//     Name = "Monument to Man",
+//     SystemHref = "/wiki/Monument%20to%20Man"
+// });
 
 List<KaydenMiller.BattleTech.Helper.Cli.System> systems;
 if (File.Exists("possible-systems.json"))
@@ -32,19 +32,42 @@ else
 }
 
 var solarSystems = new List<SolarSystem>();
-try
+if (File.Exists("solar-systems.json"))
 {
-    foreach (var system in systems)
+    Console.WriteLine("UPDATING SOLAR SYSTEMS");
+    var file = File.ReadAllText("solar-systems.json");
+    solarSystems = JsonSerializer.Deserialize<List<SolarSystem>>(file) ?? [];
+
+    List<SolarSystem> systemDetailsList = [];
+    foreach (var solarSystem in solarSystems)
     {
-        var systemDetails = await page.SearchSolarSystem(system);
-        solarSystems.Add(systemDetails);
+        solarSystem.Metadata?.TryAdd("SystemName", solarSystem.Name);
+        solarSystem.Metadata?.TryAdd("SystemWikiUrl", solarSystem.WikiUrl ?? "");
+        var systemDetails = PlanetSearch.ParseFromDictionary(solarSystem.Metadata ?? []);
+        systemDetailsList.Add(systemDetails);
+    }
+    
+    var json = JsonSerializer.Serialize(systemDetailsList, new JsonSerializerOptions()
+    {
+        WriteIndented = true
+    });
+    File.WriteAllText("solar-systems-update.json", json);
+    Console.WriteLine("DONE");
+}
+else
+{
+    try
+    {
+        foreach (var system in systems)
+        {
+            var systemDetailsDictionary = await page.SearchSolarSystem(system);
+            var systemDetails = PlanetSearch.ParseFromDictionary(systemDetailsDictionary);
+            solarSystems.Add(systemDetails);
+        }
+    }
+    finally
+    {
+        var json = JsonSerializer.Serialize(solarSystems);
+        File.WriteAllText("solar-systems.json", json);
     }
 }
-finally
-{
-    var json = JsonSerializer.Serialize(solarSystems);
-    File.WriteAllText("solar-systems.json", json); 
-}
-
-
-
